feat: add roles list and user tab
