+++ webgui/webapp/src/app/pages/admin-page/components/admin-roles-tab/admin-roles-tab.component.scss
@@ -1,0 +1,24 @@
+.roles-container {
+  position: relative;
+}
+
+.cu-header {
+  display: flex;
+  align-items: center;
+  margin-bottom: 1.5rem;
+
+  h2 {
+    margin: 0;
+    font-size: 1.5rem;
+    font-weight: 600;
+  }
+}
+
+.cu-btn-container {
+  display: flex;
+  gap: 0.5rem;
+}
+
+table {
+  margin-top: 1rem;
+}
