+++ webgui/webapp/src/app/pages/admin-page/components/admin-users-tab/admin-users-tab.component.scss
@@ -1,0 +1,43 @@
+:host ::ng-deep {
+  .mat-mdc-form-field-error-wrapper {
+    padding: 0 !important;
+  }
+}
+
+:host ::ng-deep .mat-sort-header-container {
+  justify-content: center;
+}
+
+.cu-header {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+  padding-bottom: 1rem;
+}
+
+.cu-btn-container {
+  display: flex;
+  gap: 1rem;
+}
+
+button mat-icon {
+  margin-right: 0.5rem;
+}
+
+form {
+  margin-bottom: 20px;
+}
+
+table {
+  width: 100%;
+  margin-top: 20px;
+}
+
+td {
+  cursor: pointer;
+}
+
+td:hover {
+  background-color: #f5f5f5;
+}
